[WIP][1035] Fixing the deep linking issue
Picking up on the previous effort to modify the Hugo url structure
to better handle content being rendered as cards (instead of the
default category > page flow), I've found that the page slug can be
locked down to start with a '#' in conjunction with the section
landings, so that the urls that appear in searches, rss feeds and
the like will stick to the desired format.

This is WIP because there's a new bug yet to be resolved, where
the anchor quick links are not working properly. The new url is
populated but the viewport does not move. Trying alternate JS
and setting Timeouts etc to resolve.